api/scripts/elastic_ops/pyproject.toml
[tool.poetry]
name = "elasticops"
version = "0.1.0"
description = "Rest of es helper scripts"
authors = ["joel <joel@jataware.com>"]
readme = "README.md"
packages = [
  {include = "copy_es_docs"},
  {include = "job_playground"},
  {include = "pop_didx_boxytas"}
]

[tool.poetry.dependencies]
python = "^3.9"
elasticsearch = "7.12.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
dpath = "^2.1.6"
numpy = "^1.26.2"
boto3 = "^1.33.10"
fuzzywuzzy = "^0.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
